// backend/src/generated-api/models/survivor-elimination.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Survivor Fantasy League - OpenAPI 3.0.3
 * This is the Survivor Fantasy League Spec for API's
 *
 * OpenAPI spec version: 1.0.11
 * Contact: pklein111697@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

 /**
 * 
 *
 * @export
 * @interface SurvivorElimination
 */
export interface SurvivorElimination {

    /**
     * @type {string}
     * @memberof SurvivorElimination
     */
    survivorId: string;

    /**
     * The day of the season when the elimination occurred.
     *
     * @type {number}
     * @memberof SurvivorElimination
     * @example 1
     */
    day: number;

    /**
     * The order of the survivor eliminated that episode.
     *
     * @type {number}
     * @memberof SurvivorElimination
     * @example 1
     */
    rank: number;
}